# [chore] [pkg/stanza] Improve performance of HashResource