… pkg/stanza/go.sum
… receiver/filelogreceiver/go.sum
… receiver/journaldreceiver/go.sum
… receiver/otlpjsonfilereceiver/go.sum
… receiver/syslogreceiver/go.sum
… receiver/tcplogreceiver/go.sum
… receiver/udplogreceiver/go.sum
… receiver/windowseventlogreceiver/go.sum