- Use github.com/cespare/xxhash instead of hash/fnv
- Apply pooling for frequently created objects
- Remove redundant `fnvHashOut`

Before:
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/adapter
BenchmarkGetResourceID-10                	 1934391	       621.6 ns/op	     160 B/op	       3 allocs/op
BenchmarkGetResourceIDSingleResource-10    	13024560	        78.83 ns/op	      24 B/op	       2 allocs/op
BenchmarkGetResourceIDComplexResource-10    	 1950621	       606.9 ns/op	     416 B/op	      12 allocs/op
BenchmarkConverter/worker_count=1-10         	       1	1667772333 ns/op	1796170832 B/op	48620116 allocs/op
BenchmarkConverter/worker_count=2-10         	       1	1248354667 ns/op	1796157056 B/op	48620030 allocs/op
BenchmarkConverter/worker_count=4-10         	       1	1168596542 ns/op	1796160080 B/op	48620015 allocs/op
BenchmarkConverter/worker_count=6-10         	       1	1139376458 ns/op	1796167296 B/op	48620037 allocs/op
BenchmarkConverter/worker_count=8-10         	       1	1197123000 ns/op	1796160552 B/op	48620023 allocs/op
PASS
```

After:
```
BenchmarkGetResourceID-10    	                 2614424	       454.7 ns/op	      24 B/op	       1 allocs/op
BenchmarkGetResourceIDSingleResource-10    	16507807	        63.31 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetResourceIDComplexResource-10    	 2004913	       587.4 ns/op	     376 B/op	      10 allocs/op
BenchmarkConverter/worker_count=1-10         	       1	1402426667 ns/op	1692160448 B/op	46620127 allocs/op
BenchmarkConverter/worker_count=2-10         	       1	1221293750 ns/op	1692162128 B/op	46620083 allocs/op
BenchmarkConverter/worker_count=4-10         	       1	1147778875 ns/op	1692162232 B/op	46620084 allocs/op
BenchmarkConverter/worker_count=6-10         	       1	1104916375 ns/op	1692152000 B/op	46620016 allocs/op
BenchmarkConverter/worker_count=8-10         	       1	1105069458 ns/op	1692144736 B/op	46620027 allocs/op
```